Do simple refactoring
